requirements.txt
# pip install -r requirements.txt

django

djangorestframework
markdown
django-filter
